Module cumulus_pallet_xcmp_queue::migration

A module that is responsible for migration of storage.

Constants

STORAGE_VERSION

The current storage version.

Functions

migrate_to_latest

Migrates the pallet storage to the most recent version, checking and setting the StorageVersion.

migrate_to_v1

Migrates QueueConfigData from v0 (without the xcmp_max_individual_weight field) to v1 (with max individual weight). Uses the Default implementation of QueueConfigData to choose a value for xcmp_max_individual_weight.