Function cumulus_pallet_xcmp_queue::pallet::dispatchables::update_resume_threshold

pub fn update_resume_threshold<T: Config>(new: u32)

Overwrites the number of pages which the queue must be reduced to before it signals that message sending may recommence after it has been suspended.

origin: Must pass Root.

new: Desired value for QueueConfigData.resume_threshold

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::update_resume_threshold.