Type Alias cumulus_pallet_xcmp_queue::pallet::Module

pub type Module<T> = Pallet<T>;

๐Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn suspend_xcm_execution(origin: OriginFor<T>) -> DispatchResult

Suspends all XCM executions for the XCMP queue, regardless of the senderโs origin.

origin: Must pass ControllerOrigin.

pub fn resume_xcm_execution(origin: OriginFor<T>) -> DispatchResult

Resumes all XCM executions for the XCMP queue.

Note that this function doesnโt change the status of the in/out bound channels.

origin: Must pass ControllerOrigin.

pub fn update_suspend_threshold( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Overwrites the number of pages which must be in the queue for the other side to be told to suspend their sending.

origin: Must pass Root.

new: Desired value for QueueConfigData.suspend_value

pub fn update_drop_threshold(origin: OriginFor<T>, new: u32) -> DispatchResult

Overwrites the number of pages which must be in the queue after which we drop any further messages from the channel.

origin: Must pass Root.

new: Desired value for QueueConfigData.drop_threshold

pub fn update_resume_threshold(origin: OriginFor<T>, new: u32) -> DispatchResult

Overwrites the number of pages which the queue must be reduced to before it signals that message sending may recommence after it has been suspended.

origin: Must pass Root.

new: Desired value for QueueConfigData.resume_threshold

impl<T: Config> Pallet<T>

pub fn on_idle_weight() -> Weight

The worst-case weight of on_idle.