Module cumulus_pallet_xcmp_queue::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

resume_xcm_execution

Resumes all XCM executions for the XCMP queue.

service_overweight

Services a single overweight XCM.

suspend_xcm_execution

Suspends all XCM executions for the XCMP queue, regardless of the sender’s origin.

update_drop_threshold

Overwrites the number of pages of messages which must be in the queue after which we drop any further messages from the channel.

update_resume_threshold

Overwrites the number of pages of messages which the queue must be reduced to before it signals that message sending may recommence after it has been suspended.

update_suspend_threshold

Overwrites the number of pages of messages which must be in the queue for the other side to be told to suspend their sending.

update_threshold_weight

Overwrites the amount of remaining weight under which we stop processing messages.

update_weight_restrict_decay

Overwrites the speed to which the available weight approaches the maximum weight. A lower number results in a faster progression. A value of 1 makes the entire weight available initially.

update_xcmp_max_individual_weight

Overwrite the maximum amount of weight any individual message may consume. Messages above this weight go into the overweight queue and may only be serviced explicitly.