cumulus_pallet_xcmp_queue::pallet
Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

InboundXcmpMessages

Inbound aggregate XCMP messages. It can only be one per ParaId/block.

InboundXcmpStatus

Status of the inbound XCMP channels.

OutboundXcmpMessages

The messages outbound in a given XCMP channel.

OutboundXcmpStatus

The non-empty XCMP channels in order of becoming non-empty, and the index of the first and last outbound message. If the two indices are equal, then it indicates an empty queue and there must be a non-Ok OutboundStatus. We assume queues grow no greater than 65535 items. Queue indices for normal messages begin at one; zero is reserved in case of the need to send a high-priority signal message this block. The bool is true if there is a signal message waiting to be sent.

Overweight

The messages that exceeded max individual message weight budget.

OverweightCount

The number of overweight messages ever recorded in Overweight. Also doubles as the next available free overweight index.

QueueConfig

The configuration which controls the dynamics of the outbound queue.

QueueSuspended

Whether or not the XCMP queue is suspended from executing incoming XCMs or not.

SignalMessages

Any signal messages waiting to be sent.