Module cumulus_pallet_xcmp_queue::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

DeliveryFeeFactor

The factor to multiply the base delivery fee by.

InboundXcmpSuspended

The suspended inbound XCMP channels. All others are not suspended.

OutboundXcmpMessages

The messages outbound in a given XCMP channel.

OutboundXcmpStatus

The non-empty XCMP channels in order of becoming non-empty, and the index of the first and last outbound message. If the two indices are equal, then it indicates an empty queue and there must be a non-Ok OutboundStatus. We assume queues grow no greater than 65535 items. Queue indices for normal messages begin at one; zero is reserved in case of the need to send a high-priority signal message this block. The bool is true if there is a signal message waiting to be sent.

QueueConfig

The configuration which controls the dynamics of the outbound queue.

QueueSuspended

Whether or not the XCMP queue is suspended from executing incoming XCMs or not.

SignalMessages

Any signal messages waiting to be sent.