Trait cumulus_pallet_xcmp_queue::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type XcmExecutor: ExecuteXcm<Self::RuntimeCall>;
    type ChannelInfo: GetChannelInfo;
    type VersionWrapper: WrapVersion;
    type ExecuteOverweightOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type ControllerOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type ControllerOriginConverter: ConvertOrigin<Self::RuntimeOrigin>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type XcmExecutor: ExecuteXcm<Self::RuntimeCall>

Something to execute an XCM message. We need this to service the XCMoXCMP queue.

type ChannelInfo: GetChannelInfo

Information on the avaialble XCMP channels.

type VersionWrapper: WrapVersion

Means of converting an Xcm into a VersionedXcm.

type ExecuteOverweightOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin that is allowed to execute overweight messages.

type ControllerOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin that is allowed to resume or suspend the XCMP queue.

type ControllerOriginConverter: ConvertOrigin<Self::RuntimeOrigin>

The conversion function used to attempt to convert an XCM MultiLocation origin to a superuser origin.

type WeightInfo: WeightInfo

The weight information of this pallet.